superpositioned fields cover missing fields (#28)
This PR fixes a bug where fields which were superpositioned with others
but did not have domains which fully engulfed the other fields were
required to be passed for a `write`.

If a field Foo partially overlaps with field Bar and the fields have
disjoint entitlement spaces, then the passing of only one of those
fields in a `write` gate will *cover* the other field, meaning it need
not be provided.

Note: It is conceivable that the unprovided field could forbid the value
that would end up being written to it when the other field is written.
This would only be a problem when the field ontologies swap, as it would
then have the invalid value. I've never seen hardware designed in this
way, but it is something to keep in mind!